removed use of useLocation for dateRowStartDate
