+++ frontend/src/routes/FacilityBooking/ViewBookingDailyView/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react'
 import styled from 'styled-components'
-import { useHistory, useLocation, useParams } from 'react-router-dom'
+import { useHistory, useParams } from 'react-router-dom'
 import { RootState } from '../../../store/types'
 import { useDispatch, useSelector } from 'react-redux'
 import LoadingSpin from '../../../components/LoadingSpin'
@@ -10,6 +10,7 @@ import TopNavBarRevamp from '../../../components/TopNavBarRevamp'
 import ButtonComponent from '../../../components/Button'
 import { setIsLoading, updateDailyView } from '../../../store/facilityBooking/action'
 import DailyViewDatesRow from '../../../components/FacilityBooking/DailyViewDatesRow'
+import { getDateRowStartDate } from '../../../common/getDateRowStartDate'
 
 const HEADER_HEIGHT = '70px'
 
@@ -51,12 +52,11 @@ export default function ViewBookingDailyView() {
   const history = useHistory()
   const dispatch = useDispatch()
   const params = useParams<{ facilityId: string }>()
-  const location = useLocation<ViewBookingLocationState>()
   const { clickedDate } = useSelector((state: RootState) => state.calendar)
   const { isLoading, selectedFacilityName } = useSelector((state: RootState) => state.facilityBooking)
 
   const selectedFacilityId = parseInt(params.facilityId)
-  const dateRowStartDate = location.state.dateRowStartDate
+  const dateRowStartDate = getDateRowStartDate(clickedDate)
 
   useEffect(() => {
     dispatch(setIsLoading(true))
@@ -76,14 +76,7 @@ export default function ViewBookingDailyView() {
           <ButtonComponent
             state="primary"
             text="Book Facility"
-            onClick={() =>
-              history.push({
-                pathname: `${PATHS.CREATE_FACILITY_BOOKING_DAILY_VIEW}/${selectedFacilityId}`,
-                state: {
-                  dateRowStartDate: dateRowStartDate,
-                },
-              })
-            }
+            onClick={() => history.push(`${PATHS.CREATE_FACILITY_BOOKING_DAILY_VIEW}/${selectedFacilityId}`)}
             size="small"
           />
         }
