+++ frontend/src/routes/FacilityBooking/ViewFacility/index.tsx
@@ -48,32 +48,8 @@ export default function ViewFacility() {
     />
   )
 
-  const getDateRowStartDate = (fullDate: Date) => {
-    const date = fullDate.getDate()
-    const day = fullDate.getDay()
-
-    if (day === 6) return date // if saturday chosen, start row on saturday instead of sunday
-
-    if (date - day < 0) {
-      // if date is on the new week of the month that doesn't start on sunday
-      const year = fullDate.getFullYear()
-      const month = fullDate.getMonth() // month index (0 = January)
-      const maxDatePrevMonth = new Date(year, month, 0).getDate() // max date of PREVIOUS month
-      return maxDatePrevMonth - (day - date) // get previous month's last sunday
-    }
-
-    return date - day // by default, date row start on sunday
-  }
-
   const onDateClick = (date: Date) => {
-    const dateRowStartDate = getDateRowStartDate(date)
-
-    history.push({
-      pathname: `${PATHS.VIEW_FACILITY_BOOKING_DAILY_VIEW}/${params.facilityId}`,
-      state: {
-        dateRowStartDate: dateRowStartDate,
-      },
-    })
+    history.push(`${PATHS.VIEW_FACILITY_BOOKING_DAILY_VIEW}/${params.facilityId}`)
   }
 
   return (
